open-ai/humanoid/requirement.txt
neat-python
gym==0.24.0
gym[box2D]
gym[mujoco]
graphviz
tqdm

